+++ WSCServer/include/hl_hashwrapper.h
@@ -224,7 +224,7 @@ class hashwrapper
 			/*
 			 * we update the context with the given text
 			 */
-			updateContext((unsigned char*) text.c_str(),text.length());
+			updateContext((unsigned char*) text.c_str(),(int)text.length());
 
 			/*
 			 * now we can close the hash process 
@@ -276,7 +276,7 @@ class hashwrapper
 			 * read the file in 1024b blocks and
 			 * update the context for every block
 			 */
-			while( (len = fread(buffer,1,1024,file)) )
+			while( (len = fread(buffer,(int)1,(int)1024,file)) )
 			{
 				updateContext(buffer, len);
 			}
